requirements.txt
Django~=4.0.4
Pillow~=9.1.0
azure-cognitiveservices-vision-face==0.6.0
django_compressor~=4.0.0
django-libsass~=0.9
django-crispy-forms~=1.14.0
django4-background-tasks~=1.2.7
psycopg2-binary~=2.9.3
msrest~=0.6.21